dev: Fix Revisions declaration in typed data module and RevisionField declaration in revisions module to prevent circular imports from occuring
