+++ starknet_py/net/models/typed_data.py
@@ -3,14 +3,22 @@ TypedDict structures for TypedData
 """
 
 import sys
+from enum import Enum
 from typing import Any, Dict, List, Optional, TypedDict
-
-from starknet_py.net.schemas.common import Revision
 
 if sys.version_info < (3, 11):
     from typing_extensions import NotRequired
 else:
     from typing import NotRequired
+
+
+class Revision(Enum):
+    """
+    Enum representing the revision of the specification to be used.
+    """
+
+    V0 = 0
+    V1 = 1
 
 
 class ParameterDict(TypedDict):
